% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleStatistics.R
\name{calculateAgeLengthKey}
\alias{calculateAgeLengthKey}
\title{Age-length key}
\usage{
calculateAgeLengthKey(
  ages,
  lenghts,
  ageCategories = NULL,
  lengthCategories = NULL
)
}
\arguments{
\item{ages}{vector with age observations for a sample}

\item{lenghts}{vector with length observations for a sample}

\item{ageCategories}{the categories that was recorded for age, if present in sample.}

\item{lengthCategories}{the categories that was recorded for length if present in sample.}
}
\value{
named matrix with proprotions of ages (rows) in each length group (columns). All rows sum to 1.
}
\description{
Calculates the proportion of ages in each length group.
}
\details{
Will count non-observed categories as zero
'ages' and 'lengths' are observations from the same sample, so that observations of the two variables are paired by the index to these vectors.
if 'ageCategories' or 'lengthCategories' is NULL, the present values will be used as the complete set of categories for the corresponding variable.
}
\examples{
 ages <- c(4,5,5,6,10)
 lengths <- c("20-30", "30-40","40-50", "30-40","40-50")
 categoriesAges <- 1:20
 categoriesLengths <- c("0-10","10-20", "20-30","30-40","40-50")
 calculateAgeLengthKey(ages, lengths, categoriesAges, categoriesLengths)
}
