% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{catchAtAge}
\alias{catchAtAge}
\title{Catch at age}
\description{
Estimate of catch at age in numbers
}
\details{
list with two memebers:
\describe{
\item{catchAtAge}{numeric() named vector with catch at age in numbers for each age group}
\item{covariance}{matrix() with covariance of catch at age between age groups}
}
}
