% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareEstimators.R
\name{checkEstimatorsExamplePopulation}
\alias{checkEstimatorsExamplePopulation}
\title{Check estimator}
\usage{
checkEstimatorsExamplePopulation(
  estimator,
  popParameter,
  iterations,
  nPSU,
  nSSU,
  population = lotteryEstimator::examplePopulation,
  PSU = "vesselId",
  SSU = "opid"
)
}
\arguments{
\item{estimator}{function that takes a single parameter (the output from \code{\link[lotteryEstimator]{sampleExamplePopulation}}) and returns an estimate (vector or scalar).}

\item{popParameter}{The value of the population parameter to compare the estimate to (vector or scalar)}

\item{iterations}{the number of iterations to run}

\item{nPSU}{parameter passed to \code{\link[lotteryEstimator]{sampleExamplePopulation}}}

\item{nSSU}{parameter passed to \code{\link[lotteryEstimator]{sampleExamplePopulation}}}

\item{population}{parameter passed to \code{\link[lotteryEstimator]{sampleExamplePopulation}}}

\item{PSU}{parameter passed to \code{\link[lotteryEstimator]{sampleExamplePopulation}}}

\item{SSU}{parameter passed to \code{\link[lotteryEstimator]{sampleExamplePopulation}}}
}
\value{
list with the following values (vector or scalar):
\describe{
\item{simulated.bias}{The difference between the arithmetric mean of estimates and 'popParameter'}
\item{simulated.relative.bias}{simulated.bias divided by 'popParameter'}
\item{mean.sq.error}{The mean squared error of estimates}
}
}
\description{
checks an estimator against a know population by resampling
}
\details{
The analysis is restricted to 2-stage sampling using \code{\link[lotteryEstimator]{sampleExamplePopulation}}. Parameters 'nPSU', 'nSSU', 'population', 'PSU' and 'SSU' are simply passed to this function.
}
\examples{
 data(examplePopulation)

 #total COD in example population
 total <- sum(examplePopulation$wholeWeightKg[examplePopulation$speciesFAO=="COD"])

 #Naive estimator treat clustered sampling as unclustered
 est <- function(sample){
    mean(sample$wholeWeightKg[sample$speciesFAO=="COD"])*
     sum(examplePopulation$speciesFAO=="COD")}

 #evaluate naive estimator
 \dontrun{checkEstimatorsExamplePopulation(est, total, 1000, 20, 1)}

 #confirm that naive estimator is correct when entire population is sampled
 checkEstimatorsExamplePopulation(est, total, 2, 72, 365)

}
