% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleStatistics.R
\name{countCategorical}
\alias{countCategorical}
\title{Count categorical observations}
\usage{
countCategorical(sample, categories = NULL)
}
\arguments{
\item{sample}{vector with sample observations for a categorical variable}

\item{categories}{the categories that was recorded if present in sample.}
}
\value{
named vector with counts in each category
}
\description{
Counts the number of observations of some category in a sample
}
\details{
Will count non-observed categories as zero
}
\examples{
 ages <- c(4,5,5,6,10)
 categories <- 1:20
 countCategorical(ages, categories)
}
