% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimateFromStrataTotals}
\alias{estimateFromStrataTotals}
\title{Estimate total}
\usage{
estimateFromStrataTotals(stratifiedCatchAtAge)
}
\arguments{
\item{stratifiedCatchAtAge}{estimates in each strata, formatted as \code{\link[lotteryEstimator]{stratifiedCatchAtAge}}}
}
\value{
estimates of total catch at age, formatted as \code{\link[lotteryEstimator]{catchAtAge}}
}
\description{
Estimates total catch at age from estiamtes in strata.
}
