% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hansenHurwitz}
\alias{hansenHurwitz}
\title{Hansen-Hurwitz estimator}
\usage{
hansenHurwitz(sampleTotals, selectionProbabilities)
}
\arguments{
\item{sampleTotals}{list() of numeric() vectors, each corresponding to a sample of the population parameters to be estimated.}

\item{selectionProbabilities}{numeric() vector of selection probabilites corresponding to the rows in 'sampleTotals'.}
}
\value{
numeric() vector containing the estimated population totals.
}
\description{
Hansen-Hurwitz estimator of the population total for a set of population parameters.
}
\details{
Implement estimation from hierarchical sampling by successive application of this estimator.
}
