% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hansenHurwitzInter}
\alias{hansenHurwitzInter}
\title{Inter-unit covariance of Hansen-Hurwitz estimator}
\usage{
hansenHurwitzInter(sampleTotals, selectionProbabilities)
}
\arguments{
\item{sampleTotals}{list() of numeric() vectors, each corresponding to a sample of the population parameters to be estimated.}

\item{selectionProbabilities}{numeric() vector of selection probabilites for the samples listed in 'sampleTotals'.}
}
\value{
matrix() representing a symmetric matrix with the estimated inter-unit covariances.
}
\description{
Estimator of the first terms of the covariance of a Hansen-Hurwitz estimate.
That is the covariance due to variation between sampling units.
}
\details{
This function estimates the covariance due to sampling variation between sampling untis.
The within-unit variance can be estimated with \code{\link[lotteryEstimator]{HansenHurwitzIntra}}.
Combine these estimators for hierarchical sampling.

For single-stage sampling, the sample totals are known, and there is no within-unit sampling variance.
}
