% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{HierarchicalCovarianceEstimator}
\alias{HierarchicalCovarianceEstimator}
\title{Hierarchical covariance estimator}
\description{
Function contract for hierarchical covariance estimators.
Hierarchical covariance estimators are estimators that depend on other estimates being made for each sampling unit.
A hierachical covariance estimator will divide the sample into subsets with data for each sampling unit
and call another function to calculate estimates from each of these subsets.
}
\details{
These functions take three arguments:
\describe{
\item{sample}{\code{\link{data.table}{data.table}} with the sample to estimate from}
\item{sampleId}{character() identifying column in 'sample' encoding the sampling units to estimate from}
\item{subEstimator}{that will provide a point estimate for each sampling unit}
\item{subCovarianceEstimator}{that will provide a covariance estimate for each sampling unit}
}
}
