% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hierarchicalHansenHurwitz}
\alias{hierarchicalHansenHurwitz}
\title{Hierarchical Hansen-Hurwitz}
\usage{
hierarchicalHansenHurwitz(
  sample,
  sampleId,
  subEstimator,
  selectionProbabilities
)
}
\arguments{
\item{sample}{\code{\link[data.table]{data.table}} with sample data}

\item{sampleId}{character() identifying the columns in 'sample' that identify the sampling units to estimate from}

\item{subEstimator}{function, function for estimating totals for each sample}

\item{selectionProbabilities}{character() indentifying the columns in 'sample' with selection probabilites for the sampling units.}
}
\value{
numeric() estimate of total
}
\description{
\code{\link[lotteryEstimator]{HierarchicalEstimator}} for applying
\code{\link[lotteryEstimator]{hansenHurwitz}} in hierarchical implementations.
}
\examples{
 numAtAgeSample <- function(sample){countCategorical(sample$age, 2:20)}
 numAtAgeHaul <- function(sample){hierarchicalHorwitzThompson(sample, "SSUid",
                                  numAtAgeSample, "SSUinclusionProb")}
 exampleSamples <- lotteryEstimator::NSSH2019
 exampleSamples$SSUinclusionProb <- exampleSamples$SSUselectionProb * exampleSamples$nSSU
 numAtAgeTotal <- hierarchicalHansenHurwitz(exampleSamples, "PSUid",
                                            numAtAgeHaul, "PSUselectionProb")
}
