% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hierarchicalHansenHurwitzCovariance}
\alias{hierarchicalHansenHurwitzCovariance}
\title{Hierarchical Hansen-Hurwitz covariance}
\usage{
hierarchicalHansenHurwitzCovariance(
  sample,
  partitionId,
  subEstimator,
  subCovarianceEstimator,
  selectionProbabilities
)
}
\arguments{
\item{sample}{\code{\link[data.table]{data.table}} with sample data}

\item{partitionId}{character() identifying the columns in 'sample' that identify the sampling units to estimate from}

\item{subEstimator}{function, function for estimating totals for each sampled unit}

\item{subCovarianceEstimator}{\code{\link[lotteryEstimator]{ParameterizedCovarianceEstimator}} for estimating covariances for each sampled unit}

\item{selectionProbabilities}{character() indentifying the columns in 'sample' with selection probabilites for the sampling units.}
}
\value{
numeric() estimate of total
}
\description{
\code{\link[lotteryEstimator]{HierarchicalCovarianceEstimator}} for estimating
covariance of \code{\link[lotteryEstimator]{hansenHurwitz}} in hierarchical implementations.
}
\examples{
 numAtAgeSample <- function(sample){countCategorical(sample$age, 2:20)}
 numAtAgeHaul <- function(sample){hierarchicalHorwitzThompsonTotals(sample, "SSUid",
                                  numAtAgeSample, "SSUinclusionProb")}
 exampleSamples <- lotteryEstimator::NSSH2019
 exampleSamples$SSUinclusionProb <- exampleSamples$SSUselectionProb * exampleSamples$nSSU
 covariance <- hierarchicalHansenHurwitzCovariance(exampleSamples, "PSUid",
                                                   numAtAgeHaul, function(x){0},
                                                   "PSUselectionProb")
}
