% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hierarchicalHansenHurwitzDomainTotals}
\alias{hierarchicalHansenHurwitzDomainTotals}
\title{Hierarchical Hansen-Hurwitz domain estimate}
\usage{
hierarchicalHansenHurwitzDomainTotals(
  sample,
  domainIds,
  subEstimator,
  domain,
  domainFraction
)
}
\arguments{
\item{sample}{\code{\link[data.table]{data.table}} with sample data}

\item{domainIds}{character() identifying the column in 'sample' that identify a partitioning of the sample into domains}

\item{subEstimator}{\code{\link[lotteryEstimator]{ParameterizedEstimator}} for estimating totals for each sampled unit}

\item{domain}{character() name of the domain to estimate for (must occur in the column 'domainIds')}

\item{domainFraction}{numeric() the fraction of the population that fall in the domain.}
}
\description{
Hierarchical estimator for estimating domain totals using Hansen-Hurwitz estimators
}
\examples{
 numAtAgeSample <- function(sample){countCategorical(sample$age, 2:20)}
 numAtAgeHaul <- function(sample){hierarchicalHorvitzThompsonTotals(sample, "SSUid",
                                  numAtAgeSample, "SSUinclusionProb")}
 exampleSamples <- lotteryEstimator::NSSH2019

 exampleSamples$SSUinclusionProb <- exampleSamples$SSUselectionProb * exampleSamples$nSSU
 numAtAgeTotal <- function(sample){hierarchicalHansenHurwitzTotals(sample, "PSUid",
                                            numAtAgeHaul, "PSUselectionProb")}

 #define som arbitrary domains for the sake of the example
 exampleSamples$domains <- cut(as.numeric(exampleSamples$PSUid), 3, labels = c("one", "two", "three"))

 #obtain domain estimate
 hierarchicalHansenHurwitzDomainTotals(exampleSamples, "domains", numAtAgeTotal, "two", 1/3)
}
