% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hierarchicalHorvitzThompsonCovariance}
\alias{hierarchicalHorvitzThompsonCovariance}
\title{Hierarchical Horvitz-Thompson covariance}
\usage{
hierarchicalHorvitzThompsonCovariance(
  sample,
  partitionId,
  subEstimator,
  subCovarianceEstimator,
  inclusionProbabilities,
  coInclusionMatrix
)
}
\arguments{
\item{sample}{\code{\link[data.table]{data.table}} with sample data}

\item{partitionId}{character() identifying the columns in 'sample' that identify the sampling units to estimate from}

\item{subEstimator}{function, function for estimating totals for each sampled unit}

\item{subCovarianceEstimator}{\code{\link[lotteryEstimator]{ParameterizedCovarianceEstimator}} for estimating covariances for each sampled unit}

\item{inclusionProbabilities}{character() indentifying the columns in 'sample' with inclusion probabilites for the sampling units.}

\item{coInclusionMatrix}{matrix() named matrix with co-inclusion probabilites, indexed with the names in the column identified by 'inclusionProbabilities'.}
}
\value{
numeric() estimate of total
}
\description{
\code{\link[lotteryEstimator]{HierarchicalCovarianceEstimator}} for estimating
covariance of \code{\link[lotteryEstimator]{horvitzThompson}} in hierarchical implementations.
}
