% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hierarchicalHorvitzThompsonDomainTotals}
\alias{hierarchicalHorvitzThompsonDomainTotals}
\title{Hierarchical Horvitz-Thompson domain estimate}
\usage{
hierarchicalHorvitzThompsonDomainTotals(
  sample,
  domainIds,
  subEstimator,
  domain
)
}
\arguments{
\item{sample}{\code{\link[data.table]{data.table}} with sample data}

\item{domainIds}{character() identifying the column in 'sample' that identify a partitioning of the sample into domains}

\item{subEstimator}{\code{\link[lotteryEstimator]{ParameterizedEstimator}} for estimating totals for each sampled unit}

\item{domain}{character() name of the domain to estimate for (must occur in the column 'domainIds')}
}
\description{
\code{\link[lotteryEstimator]{HierarchicalEstimator}} for estimating domain totals using Horvitz-Thompson estimators.
}
\details{
Note that no corresponding function for Horvitz-Thompson domain variance estimation is defined.
}
