% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hierarchicalHorwitzThompson}
\alias{hierarchicalHorwitzThompson}
\title{Hierarchical Howritz Thompson}
\usage{
hierarchicalHorwitzThompson(
  sample,
  sampleId,
  subEstimator,
  inclusionProbabilities
)
}
\arguments{
\item{sample}{\code{\link[data.table]{data.table}} with sample data}

\item{sampleId}{character() identifying the columns in 'sample' that identify the sampling units to estimate from}

\item{subEstimator}{function, function for estimating totals for each sample}

\item{inclusionProbabilities}{character() indentifying the columns in 'sample' with inclusion probabilites for the sampling units.}
}
\value{
numeric() estimate of total
}
\description{
\code{\link[lotteryEstimator]{HierarchicalEstimator}} for applying
\code{\link[lotteryEstimator]{horvitzThompson}} in hierarchical implementations.
}
\details{
Example provided in the documentation for \code{\link[lotteryEstimator]{hierarchicalHansenHurwitz}}
}
