% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hierarchicalHorwitzThompsonTotals}
\alias{hierarchicalHorwitzThompsonTotals}
\title{Hierarchical Howritz Thompson}
\usage{
hierarchicalHorwitzThompsonTotals(
  sample,
  partitionId,
  subEstimator,
  inclusionProbabilities
)
}
\arguments{
\item{sample}{\code{\link[data.table]{data.table}} with sample data}

\item{partitionId}{character() identifying the columns in 'sample' that identify the sampling units to estimate from}

\item{subEstimator}{\code{\link[lotteryEstimator]{ParameterizedEstimator}} for estimating totals for each sample}

\item{inclusionProbabilities}{character() indentifying the columns in 'sample' with inclusion probabilites for the sampling units.}
}
\value{
numeric() estimate of total
}
\description{
\code{\link[lotteryEstimator]{HierarchicalEstimator}} for applying
\code{\link[lotteryEstimator]{horvitzThompson}} in hierarchical implementations.
}
\details{
Example provided in the documentation for \code{\link[lotteryEstimator]{hierarchicalHansenHurwitzTotals}}
}
