% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{hierarchicalStratifiedTotals}
\alias{hierarchicalStratifiedTotals}
\title{Hierarchical stratified estimate}
\usage{
hierarchicalStratifiedTotals(sample, partitionId, subEstimator)
}
\arguments{
\item{sample}{\code{\link[data.table]{data.table}} with sample data}

\item{partitionId}{character() identifying the columns in 'sample' that identify the strata to estimate from}

\item{subEstimator}{\code{\link[lotteryEstimator]{ParameterizedEstimator}} for estimating totals for each strata}
}
\value{
vector of parameter estimates
}
\description{
\code{\link[lotteryEstimator]{HierarchicalEstimator}} for estimating totals for a stratified sample
}
