% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{horvitzThompson}
\alias{horvitzThompson}
\title{Horvitz-Thompson estimator}
\usage{
horvitzThompson(sampleTotals, inclusionProbabilities, tolerance = 1e-09)
}
\arguments{
\item{sampleTotals}{list() of numeric() vectors, each corresponding to a sample of the population parameters to be estimated.}

\item{inclusionProbabilities}{numeric() vector of inclusion probabilites corresponding to the rows in 'sampleTotals'.}

\item{tolerance}{numeric() tolerance for checks on sum of probabilities}
}
\value{
numeric() vector containing the estimated population totals.
}
\description{
Horvitz-Thompson estimator of the population total for a set of population parameters.
}
\details{
Implement estimation from hierarchical sampling by successive application of this estimator.
}
