% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{horvitzThompsonCovariance}
\alias{horvitzThompsonCovariance}
\title{Covariance of Horvitz-Thompson estimator}
\usage{
horvitzThompsonCovariance(
  sampleTotals,
  inclusionProbabilities,
  coInclusionProbabilities
)
}
\arguments{
\item{sampleTotals}{list() of numeric() vectors, each corresponding to a sample of the population parameters to be estimated.}

\item{inclusionProbabilities}{numeric() vector of inclusion probabilites for the samples whoose covariance are listed in 'sampleCovariances'.}

\item{coInclusionProbabilities}{matrix()  of co-inclusion probabilites (joint-inclusionprobabilities) for the samples whoose covariance are listed in 'sampleCovariances'.}
}
\value{
matrix() representing a symmetric matrix with the estimated covariances.
}
\description{
Estimator of the covariance of a Horvitz-Thompson estimate for single stage sampling without replacement.
That is the covariance due to variation between sampling units.
}
\details{
For multi-stage sampling, this function estimates the covariance due to sampling variation between sampling untis.
The within-unit variance can be estimated with \code{\link[lotteryEstimator]{horvitzThompsonIntra}}.
Combine these estimators for total covariance in hierarchical sampling.
}
