% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{horvitzThompsonIntra}
\alias{horvitzThompsonIntra}
\title{Intra-unit covariance of Horvitz-Thompson estimator}
\usage{
horvitzThompsonIntra(sampleCovariances, inclusionProbabilities)
}
\arguments{
\item{sampleCovariances}{list() of matrix() matrices, with the covariances of sample estimates.}

\item{inclusionProbabilities}{numeric() vector of inclusion probabilites for the samples whoose covariance are listed in 'sampleCovariances'.}
}
\value{
data.table() representing a symmetric matrix with the estimated covariances.

matrix() representing a symmetric matrix with the estimated covariances. intra-unit covariances.
}
\description{
Estimator of the last term of the covariance of a Horvitz-Thompson estimate for multi-stage sampling without replacement.
That is the covariance due to variation within sampling units.
}
\details{
This function estimates the covariance due to sampling variation within sampling untis.
The between-unit variance can be estimated with \code{\link[lotteryEstimator]{horvitzThompsonCovariance}}.
Combine these estimators for hierarchical sampling.

For single-stage sampling, the sample totals are known, and there is no within-unit sampling variance.
}
