% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSSH2019-datadoc.R
\docType{data}
\name{NSSH2019}
\alias{NSSH2019}
\title{Example data - Norwegian Spring Spawning Herring sampled in 2019}
\format{\code{\link[data.table]{data.table}} one row for each measured fish and with columns:
\describe{
\item{PSUid}{Identifier for the primary sampling unit (PSU) the fish was sampled from}
\item{SSUid}{Identifier for the secondary sampling unit (SSU) the fish was sampled from}
\item{PSUselectionProb}{Selection probability for the PSU}
\item{SSUselectionProb}{Selection probability for the SSU}
\item{age}{The age determied for the fish}
\item{length}{The length determined for the fish}
}}
\usage{
data(NSSH2019)
}
\description{
Example data - Norwegian Spring Spawning Herring sampled in 2019
}
\examples{
data(NSSH2019)

}
\keyword{datasets}
