% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{ParameterizedCovarianceEstimator}
\alias{ParameterizedCovarianceEstimator}
\title{Parameterized covariance estimator}
\description{
Function contract for parameterized covariance estimators
Parameterized covariance estimators are functions that map a data table to a covariance estimate, with all other parameters fixed.
}
\details{
Parameterized covariance estimators take a single argument 'sample' which is a \code{\link[data.table]{data.table}} with sample data

Parameterized covariance estimators return a named matrix with estimates of covariances between parameters

Parameterized covariance estimators can be obtained from \code{\link[lotteryEstimator]{HierarchicalCovarianceEstimator}}
by fixing parameters, in the same manner that this is done for \code{\link[lotteryEstimator]{ParameterizedEstimator}}
}
