% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{ParameterizedEstimator}
\alias{ParameterizedEstimator}
\title{Parameterized estimator}
\description{
Function contract for parameterized estimators
Parameterized estimators are functions that map a data frame to an estimate, with all other parameters fixed.
}
\details{
Parameterized estimators take a single argument 'sample' which is a \code{\link[data.table]{data.table}} with sample data

Parameterized estimators return a named numeric vector with parameter estimates
}
\examples{
 # A parameterized estimator can be obtained from
 # \code{\link[lotteryEstimator]{HierarchicalEstimator}}
 # by fixing parameters.
 # For example a parametierized Horwitz-Thomposon estimator:
 numAtAgeSample <- function(sample){countCategorical(sample$age, 2:20)}
 numAtAgeHaul <- function(sample){hierarchicalHorvitzThompsonTotals(sample, "SSUid",
                                  numAtAgeSample, "SSUinclusionProb")}


}
