% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poissionSample}
\alias{poissionSample}
\title{Possion sampling}
\usage{
poissionSample(population, selectionProbabilities, sampleSize)
}
\arguments{
\item{population}{numeric: vector representing the population to sample from}

\item{selectionProbabilities}{numeric: vector representing the inclusion probabilites of the populatin}

\item{sampleSize}{numeric: expected sample size}
}
\value{
numeric a subset of 'population'
}
\description{
Perform Poisson sampling
}
\details{
Sampling is performed on vectors. For sampling a data frame, perform sampling on indecies.

Note the definition of 'sampleSize'. In poisson sampling the sample size is not fixed and may vary from draw to draw.
}
\examples{
 data("longlinerPopulation")
 codset <- longlinerPopulation[longlinerPopulation$speciesFAO == "COD",]

 #select cod catches with inclusion probabilites proportional to their weight
 codselection <- codset[poissionSample(1:nrow(codset), codset$wholeWeightKg/sum(codset$wholeWeightKg), 100)]
 nrow(codselection)
}
