% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poissonInclusionProbability}
\alias{poissonInclusionProbability}
\title{Poisson inclusion probability}
\usage{
poissonInclusionProbability(selectionProbability, sampleSize)
}
\arguments{
\item{selectionProbability}{numeric: selection probability for sample}

\item{sampleSize}{numeric: expected sample size}
}
\description{
calculates inclusion probability for Poisson sampling, based on selection probability and expected sample size.
}
\examples{
 poissonInclusionProbability(.005,300)
 poissonInclusionProbability(c(.0005,0.0001),300)

 # selection proportional to fraction of total in sample
 # e.g. catch relative to quota
 inSample <- 200
 total <- 400000
 sampleSize <- 300
 poissonInclusionProbability(inSample/total,sampleSize)

}
