% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poissonJointInclusionProbabilityMatrix}
\alias{poissonJointInclusionProbabilityMatrix}
\title{Poisson joint inclusion probability}
\usage{
poissonJointInclusionProbabilityMatrix(inclusionProbabilities)
}
\arguments{
\item{inclusionProbabilities}{inclusion probabilities for a sample}
}
\value{
the pairwise joint inclusion probability matrix
}
\description{
calculates the pariwise joint inclusion probability for Poisson sampling, based on respective inclusion probabilities
}
