% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poissonSelectionProbability}
\alias{poissonSelectionProbability}
\title{Poisson selection probability}
\usage{
poissonSelectionProbability(inclusionProbability, sampleSize)
}
\arguments{
\item{inclusionProbability}{numeric: inclusion probability for sample}

\item{sampleSize}{numeric: expected sample size}
}
\description{
calculates selection probability for Poisson sampling, based on inclusion probability and expected sample size.
}
\examples{
 # selection proportional to fraction of total in sample
 # e.g. catch relative to quota
 inSample <- 200
 total <- 400000
 sampleSize <- 300
 inclusionProbability <- poissonInclusionProbability(inSample/total,sampleSize)
 selectionProbability <- poissonSelectionProbability(inclusionProbability, sampleSize)

}
