% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleStatistics.R
\name{proportionCategorical}
\alias{proportionCategorical}
\title{Proportions of categorical observations}
\usage{
proportionCategorical(sample, categories = NULL)
}
\arguments{
\item{sample}{vector with sample observations for a categorical variable}

\item{categories}{the categories that was recorded if present in sample.}
}
\value{
named vector with counts in each category
}
\description{
Calculates the proportion of observations of some category in a sample
}
\details{
Will count non-observed categories as zero
}
\examples{
 ages <- c(4,5,5,6,10)
 categories <- 1:20
 proportionCategorical(ages, categories)
}
