% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{randomNonResponseCorrection}
\alias{randomNonResponseCorrection}
\title{Random sample loss correction}
\usage{
randomNonResponseCorrection(
  inclusionProbabilities,
  selected,
  sampled = length(inclusionProbabilities)
)
}
\arguments{
\item{selected}{The number of sampling units that was selected}

\item{sampled}{The number of sampling units that was actually sampled and recorded}

\item{inclusionProbability}{numeric: vector inclusion probabilites for sampled sampling units.}
}
\value{
numeric: vector of corrected inclusion probabilites.
}
\description{
Corrects inclusion probabilities due to random loss of samples.
Apprioriate if selected sampling units are not sampled due to random inteference.
The random interference should affect different selections with equal probability
(not related to the fact that the selections may have been made with unequal probability).
}
\examples{
 # correct a vector of inclusion probabilities
 # 4 samples selected, two missing
 randomNonResponseCorrection(c(.4,.1), 4)

 # correct a single inclusion probability
 # 100 samples selected. 20 missing
 randomNonResponseCorrection(.1, 100, 80)

}
