% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonresponse.R
\name{randomNonResponseCorrectionFactor}
\alias{randomNonResponseCorrectionFactor}
\title{Random sample loss correction}
\usage{
randomNonResponseCorrectionFactor(
  selectedInclusionProbabilities,
  sampledInclusionProbabilities
)
}
\arguments{
\item{selectedInclusionProbabilities}{numeric: vector with inclusion probabilites for selected sampling units.}

\item{sampledInclusionProbabilities}{numeric: vector with inclusion probabilites for sampled sampling units ('selectedInclusionProbabilities' excluding non-response).}
}
\value{
scalar for correcting inclusionprobabilites
}
\description{
Calulcates factor for correcting inclusion probabilities due to random loss of samples.
Appropriate if selected sampling units are not sampled due to random inteference.
The random interference should affect different selections with equal probability
(not related to the fact that the selections may have been made with unequal probability).
}
\examples{
 # correct a vector of inclusion probabilities
 # 4 samples selected, two missing
 inclusionProbabilities <- c(.4,.1,.2,.1)
 nonresponse <- c(F,F,T,T)
 sampled <- inclusionProbabilities[!nonresponse]
 correctionFactor <- randomNonResponseCorrectionFactor(inclusionProbabilities, sampled)
 correctedInclusionProbabilites <- sampled * correctionFactor

 sampleTotals <- c(3,4)

 #estimate
 preEst <- horvitzThompson(sampleTotals, correctedInclusionProbabilites)

 #correct post-estimate
 postEst <- horvitzThompson(sampleTotals, inclusionProbabilities[!nonresponse])/correctionFactor

 preEst - postEst

}
