% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonresponse.R
\name{randomNonResponseCorrectionSE}
\alias{randomNonResponseCorrectionSE}
\title{Correct total random nonrespons}
\usage{
randomNonResponseCorrectionSE(respSE, responseRate)
}
\arguments{
\item{respSE}{Estimate of standard error of total based on respondents}

\item{responseRate}{the response rate.}
}
\value{
Estimate of the standard error of the total, corrected for random nonresponse
}
\description{
Horvitz-Thompson estimate of a standard error for random non-response, assuming known response rate.
}
\examples{
 #survey with an estimated SE of 30, and a responserate of .7
 randomNonResponseCorrectionSE(30, .7)
}
