% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonresponse.R
\name{randomNonResponseCorrectionTotal}
\alias{randomNonResponseCorrectionTotal}
\title{Correct total random nonrespons}
\usage{
randomNonResponseCorrectionTotal(respTotal, responseRate)
}
\arguments{
\item{respTotal}{Estimate of total based on respondents}

\item{responseRate}{the response rate.}
}
\value{
Estimate of the total, corrected for random nonresponse
}
\description{
Horvitz-Thompson estimate of a total for random non-response, assuming known response rate
}
\examples{
 #survey with an estimated total of 55, 10 respondents and a responserate of .7
 randomNonResponseCorrectionTotal(55, 10, .7)
}
