% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareEstimators.R
\name{sampleExamplePopulation}
\alias{sampleExamplePopulation}
\title{Sample example population}
\usage{
sampleExamplePopulation(
  nPSU,
  nSSU,
  population = lotteryEstimator::longlinerPopulation,
  PSU = "vesselId",
  SSU = "opid"
)
}
\arguments{
\item{nPSU}{the number of primary sampling units to sample}

\item{nSSU}{the maximal number of secondary sampling untis to sample}

\item{population}{\code{\link[data.table]{data.table}} with the entire popoulation to sample. The population must be organised in clusters for two stage sampling, identfied by 'PSU' and 'SSU'.}

\item{PSU}{character() specifying the column in 'population' that identifies the primary sampling units}

\item{SSU}{character() specifying the column in 'population' that identifies the secondary sampling units}
}
\value{
\code{\link[data.table]{data.table}} formatted as 'population' containing only the sampled elements, and the following additional columns:
\describe{
\item{Npsu}{total number of PSUs in the population}
\item{Nssu}{total number of SSUs within the given PSU in the population}
\item{nPSU}{total number of PSUs selected}
\item{nSSU}{total number of SSUs selected withing the given PSU.}
\item{ipPSU}{inclusion probability of PSU}
\item{spPSU}{selection probability of PSU}
\item{ipSSU}{inclusion probability of SSU, given that PSU is selected}
\item{spSSU}{selection probability of SSU, given that PSU is selected}
}
}
\description{
Sample example population
}
\details{
Performs two-stage sampling without replacement from an example population
}
\examples{
 sample <- sampleExamplePopulation(10,10)
}
