% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{sampleIndecies}
\alias{sampleIndecies}
\title{Sample indecies}
\usage{
sampleIndecies(n, maxindex, replacement = T, popsize = NA)
}
\arguments{
\item{n}{integer() the number of indecies to select}

\item{maxindex}{integer() defines the range of indecies to sample (1:'maxindex')}

\item{replacement}{logical() whether to sample indecies with replacement}

\item{popsize}{integer() size of the pseudopopulation to be sampled. May be NA if replacement is TRUE.}

\item{return}{integer() vector with sampled integers in the range 1:'maxindex'}
}
\description{
samples integers for indexing data tables in resampling.
}
\details{
Sampling without replacement is mimicked by sampling from a pseudopopulation.
Effectively the integers up to an including 'maxindex' are replicated
as many times as necessary to reach the size of the population.
Any remaineder is filled in with number selected without replacement from the range 1:'maxindex'.
This psudopopulation is the sampled wihtout replacement for the final result.
}
