% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs.R
\name{srsworJointInclusionProbabilityMatrix}
\alias{srsworJointInclusionProbabilityMatrix}
\title{SRS joint inclusion probability}
\usage{
srsworJointInclusionProbabilityMatrix(n, N)
}
\arguments{
\item{n}{sample size}

\item{N}{population size}
}
\value{
the pairwise joint inclusion probability matrix (n X n)
}
\description{
calculates the pariwise joint inclusion probability for simple random sampling with replacement, based on sample size and population size
}
\details{
For more convenient incorporation in generic estimators, the result is return as a n X n matrix with the joint inclusion probability repeated
}
