% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{stratifiedCatchAtAge}
\alias{stratifiedCatchAtAge}
\title{Stratified catch at age}
\description{
Stratified Estimate of catch at age in numbers
}
\details{
list with three members:
\describe{
\item{estimates}{named list of \code{\link[lotteryEstimator]{catchAtAge}} estimates, with names identifying strata the estimate was done for.}
}
}
