% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{twoStageNsshEstimator}
\alias{twoStageNsshEstimator}
\title{Estimator for NSSH}
\usage{
twoStageNsshEstimator(samples, minAge = 1, maxAge = 20)
}
\arguments{
\item{samples}{samples formatted as \code{\link[lotteryEstimator]{NSSH2019}}}

\item{minAge}{lowest age to produce estimate for}

\item{maxAge}{highest age to produce estimate for}
}
\value{
estimates, formatted as \code{\link[lotteryEstimator]{catchAtAge}}
}
\description{
Estimator for 2-stage lottery sampling, assuming with replacement haul-selection and estimating intra-haul variance from sample covariances.
}
\details{
This estimator is applicable for two stage sampling based on the lottery system.
PSU: haul, assumed UEPWR, Hansen-Hurwitz estimator
SSU: bucket, assumed SRSWOR, Horvitz-Thompson estimator
}
\examples{
 data(NSSH2019)
 estimates <- simpleNsshEstimator(NSSH2019, 1,20)
}
