% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{twoStageNsshEstimatorBootstrapped}
\alias{twoStageNsshEstimatorBootstrapped}
\title{Estimator for NSSH}
\usage{
twoStageNsshEstimatorBootstrapped(
  samples,
  minAge = 1,
  maxAge = 20,
  iterations = 10
)
}
\arguments{
\item{samples}{samples formatted as \code{\link[lotteryEstimator]{NSSH2019}}}

\item{minAge}{lowest age to produce estimate for}

\item{maxAge}{highest age to produce estimate for}
}
\value{
estimates, formatted as \code{\link[lotteryEstimator]{catchAtAge}}
}
\description{
Estimator for 2-stage lottery sampling, assuming with replacement haul-selection and estimating intra-haul variance from sample covariances.
}
\details{
As \code{\link[lotteryEstimator]{twoStageNsshEstimator}}, but with intra-haul variance bootstrapped
by selection without replacement.
}
\examples{
 data(NSSH2019)
 estimates <- twoStageNsshEstimatorBootstrapped(NSSH2019, 1,20)
}
