% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{HierarchicalEstimator}
\alias{HierarchicalEstimator}
\title{Hierarchical estimator}
\description{
Function contract for hierarchical estimators.
}
\details{
The contract specifies what the first argument and the return value.
Fix other arguments in order to form a \code{\link[lotteryEstimator]{ParameterizedEstimator}}

Hierarchical estimators are estimators that depend on other estimates being made for partitions of the sample.
Such partitions can be lower level sampling units, clusters or strata.
A hierachical estimator will divide the sample into subsets with data for each partition
and call another function to calculate estimates from each of these subsets.

These functions take one or more arguments:
\describe{
\item{sample}{\code{\link{data.table}{data.table}} with the sample to estimate from}
\item{...}{other arguments needed to calculate estimate}
}

A hierarchical estimators return a vector of estimated parameters.
}
